Oppgraderer til Spring Boot 3.
- Oppgraderer gradle wrapper.
- Oppgraderer plugins.
- Oppgraderer pakker.
- Migrerer fra javax til jakarta.
- Erstatter DefaultWebMvcTagsProvider med DefaultServerRequestObservationConvention.
- Erstatter Zalando Problem med Spring ProblemDetail.
- Erstatter Hibernate TypeDefs med Hibernate JdbcTypeCode.JSON for jsonb typer.
- Erstatter Jetty med Tomcat til Jetty støtter Servlet 6.0
